requirements.txt
selenium == 4.25.0
pytest == 8.3.3
allure-pytest == 2.13.5
seletools == 1.5.0
